+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.4.8
 * Removed direct GO mock replacement
 * Added mock support for custom vegetation
+* Added mock support for DropTable structs
 
 ## Version 2.4.7
 * Fixed mock resolving of certain components (piece place effects for example)